steamdeck/deck-reader/requirements.txt
piper-tts
sounddevice
numpy
pytesseract
Pillow
